+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
  - Performance improvements for JsonSerialBuffer
+ - XP3 Changes to work with new SPlusShim paradigm
 
 ## [8.0.0] - 2019-01-10
 ### Added