+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - Performance improvements for JsonSerialBuffer
  - XP3 Changes to work with new SPlusShim paradigm
  - XP3 NullRef fixes
+ - SigCache JSON optimizations to better handle large amounts of data
 
 ## [8.0.0] - 2019-01-10
 ### Added